+++ src/plugins/contentRenderer/cssDivs/body/cssDivsBodyContentRenderer.component.ts
@@ -71,7 +71,7 @@ export class CssDivsBodyContentRendererComponent<TData> extends BodyContentRende
     //######################### constructor #########################
     constructor(pluginElement: ElementRef,
                 changeDetector: ChangeDetectorRef,
-                @Inject(GRID_PLUGIN_INSTANCES) gridPlugins: GridPluginInstances,
+                @Inject(GRID_PLUGIN_INSTANCES) @Optional() gridPlugins: GridPluginInstances,
                 @Inject(BODY_CONTENT_RENDERER_OPTIONS) @Optional() options: CssDivsBodyContentRendererOptions)
     {
         super(pluginElement, changeDetector, gridPlugins);